+++ CHANGELOG.md
@@ -4,7 +4,29 @@
 
 # Changelog
 
-## Latest Version - v4.1.13 (08/19/25)
+## Latest Version - v4.1.14 (10/16/25)
+- [Globalpayments-js][4.1.14]  Express Pay rollback
+- [Globalpayments-js][4.1.14]  Include Usage Mode for Network Token
+
+## ***Unreleased Stuff***
+- [Globalpayments-3ds][?] Fixed issue with 3DS functionality in sandbox environment
+
+## v4.1.18 (*Unreleased to Prod*)
+- [Globalpayments-js][4.1.18]  Express pay SIT Integration issues for cert deployment
+
+## v4.1.17 (*Unreleased to Prod*)
+- [Globalpayments-js][4.1.17]  Express pay SIT Integration issues for cert deployment
+
+## Latest Version - v4.1.16 (*Unreleased to Prod*)
+- [Globalpayments-js][4.1.16]  Express pay SIT Integration issues for cert deployment
+
+## v4.1.15 (08/22/25) (*Unreleased to Prod*)
+- [Globalpayments-js][4.1.15]  Express pay SIT Integration issues for cert deployment
+
+## v4.1.14 (08/20/25)
+- [Globalpayments-js][4.1.14]  Express pay changes revert
+
+## v4.1.13 (08/19/25)
 - [Globalpayments-js][4.1.13]  Added Express pay Feature
 
 ## v4.1.12 (07/29/25)